model big ints

+++ teams/models.py
@@ -8,8 +8,8 @@ class Team(models.Model):
 		return self.name
 
 class Player(models.Model):
-	riot_id = models.IntegerField()
-	account_id = models.IntegerField()
+	riot_id = models.BigIntegerField()
+	account_id = models.BigIntegerField()
 	name = models.CharField(max_length=200)
 	profile_icon_id = models.IntegerField()
 	revision_date = models.IntegerField()
@@ -21,38 +21,38 @@ class Player(models.Model):
 
 class FlexMatch(models.Model):
 	player = models.ForeignKey(Player, on_delete=models.CASCADE)
-	gameId = models.IntegerField()
+	gameId = models.BigIntegerField()
 	champion = models.IntegerField()
 	season = models.IntegerField()
-	timestamp = models.IntegerField()
+	timestamp = models.BigIntegerField()
 	role = models.CharField(max_length=200)
 	lane = models.CharField(max_length=200)
 	game_version = models.CharField(max_length=200)
 	win = models.BooleanField()
-	top = models.IntegerField(default=0)
-	mid = models.IntegerField(default=0)
-	jun = models.IntegerField(default=0)
-	adc = models.IntegerField(default=0)
-	sup = models.IntegerField(default=0)
+	top = models.BigIntegerField(default=0)
+	mid = models.BigIntegerField(default=0)
+	jun = models.BigIntegerField(default=0)
+	adc = models.BigIntegerField(default=0)
+	sup = models.BigIntegerField(default=0)
 
 	def __str__(self):
 		return f"{self.player} {self.gameId}"
-		
+
 class SoloMatch(models.Model):
 	player = models.ForeignKey(Player, on_delete=models.CASCADE)
-	gameId = models.IntegerField()
+	gameId = models.BigIntegerField()
 	champion = models.IntegerField()
 	season = models.IntegerField()
-	timestamp = models.IntegerField()
+	timestamp = models.BigIntegerField()
 	role = models.CharField(max_length=200)
 	lane = models.CharField(max_length=200)
 	game_version = models.CharField(max_length=200)
 	win = models.BooleanField()
-	top = models.IntegerField(default=0)
-	mid = models.IntegerField(default=0)
-	jun = models.IntegerField(default=0)
-	adc = models.IntegerField(default=0)
-	sup = models.IntegerField(default=0)
+	top = models.BigIntegerField(default=0)
+	mid = models.BigIntegerField(default=0)
+	jun = models.BigIntegerField(default=0)
+	adc = models.BigIntegerField(default=0)
+	sup = models.BigIntegerField(default=0)
 
 	def __str__(self):
 		return f"{self.player} {self.gameId}"
